requirements.txt
gspread==6.1.4
pandas
google-cloud-bigquery
dbt-bigquery==1.5.3 